// Bottleneko.ScriptRuntime/neko/internal/api/bottleneko.gen.ts
export type Packet
    = AuthenticatePacket
        | MailPacket
        | RequestPacket
        | ResponsePacket;


export type ClientType
    = 'Api'
        | 'Script';
export const ClientTypeValues : ClientType[] = [
    'Api',
    'Script',
];

export interface AuthenticatePacket {
    $type: 'Authenticate';
    clientType: ClientType;
    accessToken: string;
}

export type RpcRequest
    = ConnectionsListRequest
        | ConnectionsGetRequest
        | ConnectionsUpdateRequest
        | ConnectionsDeleteRequest
        | ConnectionsAddRequest
        | ConnectionsGetAttachmentUrlRequest
        | ConnectionsStartRequest
        | ConnectionsRestartRequest
        | ConnectionsStopRequest
        | ConnectionsTestRequest
        | LoggingLogRequest
        | LoggingSubscribeRequest
        | LoggingUnsubscribeRequest
        | MessagesSubscribeRequest
        | MessagesUnsubscribeRequest
        | ScriptGetIdRequest
        | ScriptGetNameRequest;

export type RpcResponse
    = ConnectionsListResponse
        | ConnectionsGetResponse
        | ConnectionsUpdateResponse
        | ConnectionsDeleteResponse
        | ConnectionsAddResponse
        | ConnectionsGetAttachmentUrlResponse
        | ConnectionsStartResponse
        | ConnectionsRestartResponse
        | ConnectionsStopResponse
        | ConnectionsTestResponse
        | LoggingLogResponse
        | LoggingSubscribeResponse
        | LoggingUnsubscribeResponse
        | MessagesSubscribeResponse
        | MessagesUnsubscribeResponse
        | ScriptGetIdResponse
        | ScriptGetNameResponse;

export interface RequestPacket {
    $type: 'Request';
    requestId: string;
    request: RpcRequest;
}

export interface ResponsePacket {
    $type: 'Response';
    requestId: string;
    result: ResponseResult;
}

export type ResponseResult
    = SuccessResult
        | ErrorResult;

export interface SuccessResult {
    $type: 'Success';
    data: RpcResponse;
}

export interface ErrorResult {
    $type: 'Error';
    code: ErrorCode;
    message: string;
    extra: object | null;
}

export interface SubscriptionId {
    id: string;
}

export type Letter
    = LogLetter
        | ChatMessageLetter;

export interface MailPacket {
    $type: 'Mail';
    subscriptionId: SubscriptionId;
    letters: Letter[];
}
// ...
export interface LogFilter {
    severities: LogSeverity[] | null;
    sourceType: LogSourceType | null;
    sourceId: string | null;
    category: string | null;
}

export interface ChatMessageLetter {
    $type: 'ChatMessage';
    content: ChatMessageDto;
}

export interface ChatMessageFilter {
    protocol: Protocol | null;
    connectionId: string | null;
}

export interface ConnectionTestResult {
    duration: number;
    extra: object | null;
}

export interface ConnectionsListRequest {
    $type: 'Connections/List';
}

export interface ConnectionsListResponse {
    $type: 'Connections/List';
    result: ConnectionDto[];
}

export interface ConnectionsGetRequest {
    $type: 'Connections/Get';
    id: string;
}

export interface ConnectionsGetResponse {
    $type: 'Connections/Get';
    result: ConnectionDto;
}

export interface ConnectionsUpdateRequest {
    $type: 'Connections/Update';
    id: string;
    name: string | null;
    config: ProtocolConfiguration | null;
    autoStart: boolean | null;
}

export interface ConnectionsUpdateResponse {
    $type: 'Connections/Update';
    result: ConnectionDto;
}

export interface ConnectionsDeleteRequest {
    $type: 'Connections/Delete';
    id: string;
}

export interface ConnectionsDeleteResponse {
    $type: 'Connections/Delete';
}

export interface ConnectionsAddRequest {
    $type: 'Connections/Add';
    name: string;
    protocol: Protocol;
    config: ProtocolConfiguration;
}

export interface ConnectionsAddResponse {
    $type: 'Connections/Add';
    result: ConnectionDto;
}

export interface ConnectionsGetAttachmentUrlRequest {
    $type: 'Connections/GetAttachmentUrl';
    id: string;
    attachmentId: string;
}

export interface ConnectionsGetAttachmentUrlResponse {
    $type: 'Connections/GetAttachmentUrl';
    result: string;
}

export interface ConnectionsStartRequest {
    $type: 'Connections/Start';
    id: string;
}

export interface ConnectionsStartResponse {
    $type: 'Connections/Start';
}

export interface ConnectionsRestartRequest {
    $type: 'Connections/Restart';
    id: string;
}

export interface ConnectionsRestartResponse {
    $type: 'Connections/Restart';
}

export interface ConnectionsStopRequest {
    $type: 'Connections/Stop';
    id: string;
}

export interface ConnectionsStopResponse {
    $type: 'Connections/Stop';
}

export interface ConnectionsTestRequest {
    $type: 'Connections/Test';
    protocol: Protocol;
    config: ProtocolConfiguration;
}

export interface ConnectionsTestResponse {
    $type: 'Connections/Test';
    result: ConnectionTestResult;
}

export interface LoggingLogRequest {
    $type: 'Logging/Log';
    severity: LogSeverity;
    message: string;
}

export interface LoggingLogResponse {
    $type: 'Logging/Log';
}

export interface LoggingSubscribeRequest {
    $type: 'Logging/Subscribe';
    filter: LogFilter;
}

export interface LoggingSubscribeResponse {
    $type: 'Logging/Subscribe';
    result: SubscriptionId;
}

export interface LoggingUnsubscribeRequest {
    $type: 'Logging/Unsubscribe';
    subscriptionId: SubscriptionId;
}

export interface LoggingUnsubscribeResponse {
    $type: 'Logging/Unsubscribe';
}

export interface MessagesSubscribeRequest {
    $type: 'Messages/Subscribe';
    filter: ChatMessageFilter;
}

export interface MessagesSubscribeResponse {
    $type: 'Messages/Subscribe';
    result: SubscriptionId;
}

export interface MessagesUnsubscribeRequest {
    $type: 'Messages/Unsubscribe';
    subscriptionId: SubscriptionId;
}

export interface MessagesUnsubscribeResponse {
    $type: 'Messages/Unsubscribe';
}

export interface ScriptGetIdRequest {
    $type: 'Script/GetId';
}

export interface ScriptGetIdResponse {
    $type: 'Script/GetId';
    result: string;
}

export interface ScriptGetNameRequest {
    $type: 'Script/GetName';
}

export interface ScriptGetNameResponse {
    $type: 'Script/GetName';
    result: string;
}

export interface DiscordProtocolConfiguration {
    $type: 'Discord';
    token: string;
    receiveEvents: boolean;
    isPresenceIntentEnabled: boolean;
    isServerMembersIntentEnabled: boolean;
    isMessageContentIntentEnabled: boolean;
    proxyId: string | null;
}

export type ProtocolConfiguration
    = DiscordProtocolConfiguration
        | TelegramProtocolConfiguration
        | TwitchProtocolConfiguration;
// ...
export type Protocol
    = 'Discord'
        | 'Telegram'
        | 'Twitch';
export const ProtocolValues : Protocol[] = [
    'Discord',
    'Telegram',
    'Twitch',
];

export interface ExtendedConnectionStatus {
    status: ConnectionStatus;
    statusChangeDelay: number;
}

export interface ConnectionDto {
    id: string;
    name: string;
    protocol: Protocol;
    autoStart: boolean;
    config: ProtocolConfiguration;
    extendedStatus: ExtendedConnectionStatus;
}
// ...
export type LogSeverity
    = 'Critical'
        | 'Error'
        | 'Warning'
        | 'Info'
        | 'Verbose'
        | 'Debug';